+++ Common/Settings.hpp
@@ -11,7 +11,7 @@ class Settings final
 {
 public:
 
-	static Settings* CreateInstance(
+	Settings(
 		const std::string serverIP,
 		uint16_t serverPort,
 		const bool bNetworkingEnabled,
@@ -25,33 +25,15 @@ public:
 		const bool bRequireRunAsAdministrator,
 		const bool bUsingDriver,
 		const std::wstring DriverSignerSubject,
-		const list<wstring> allowedParents,
-		const bool enableLogging,
+		list<wstring> allowedParents,
+		const bool bEnableLogging,
 		const string logFileName)
+		: serverIP(serverIP), serverPort(serverPort), bNetworkingEnabled(bNetworkingEnabled), bEnforceSecureBoot(bEnforceSecureBoot), bEnforceDSE(bEnforceDSE),
+		bEnforceNoKDbg(bEnforceNoKDbg), bUseAntiDebugging(bUseAntiDebugging), bCheckIntegrity(bCheckIntegrity), bCheckThreads(bCheckThreads), bCheckHypervisor(bCheckHypervisor),
+		bRequireRunAsAdministrator(bRequireRunAsAdministrator), bUsingDriver(bUsingDriver), DriverSignerSubject(DriverSignerSubject), allowedParents(allowedParents), bEnableLogging(bEnableLogging), logFileName(logFileName)
 	{
-		if (!Instance)
-		{
-			Instance = new Settings(
-				serverIP,
-				serverPort,
-				bNetworkingEnabled, 
-				bEnforceSecureBoot, 
-				bEnforceDSE, 
-				bEnforceNoKDbg, 
-				bUseAntiDebugging, 
-				bCheckIntegrity, 
-				bCheckThreads, 
-				bCheckHypervisor, 
-				bRequireRunAsAdministrator,
-				bUsingDriver,
-				DriverSignerSubject,
-				allowedParents,
-				enableLogging,
-				logFileName
-			);
-		}
-
-		return Instance;
+		Logger::enableLogging = bEnableLogging; //put this line here to be as early as possible.
+		Logger::logFileName = logFileName;
 	}
 
 	Settings(const Settings&) = delete; //prevent copying
@@ -82,39 +64,7 @@ public:
 	wstring GetKMDriverName() const { return this->KMDriverName; }
 	wstring GetKMDriverPath() const { return this->KMDriverPath; }
 	wstring GetKMDriverSignee() const { return this->KMDriverSignee; }
-
-	Settings(
-		const std::string serverIP,
-		uint16_t serverPort,
-		const bool bNetworkingEnabled,
-		const bool bEnforceSecureBoot,
-		const bool bEnforceDSE,
-		const bool bEnforceNoKDbg,
-		const bool bUseAntiDebugging,
-		const bool bCheckIntegrity,
-		const bool bCheckThreads,
-		const bool bCheckHypervisor,
-		const bool bRequireRunAsAdministrator,
-		const bool bUsingDriver,
-		const std::wstring DriverSignerSubject,
-		list<wstring> allowedParents,
-		const bool bEnableLogging,
-		const string logFileName)
-		: serverIP(serverIP), serverPort(serverPort), bNetworkingEnabled(bNetworkingEnabled), bEnforceSecureBoot(bEnforceSecureBoot), bEnforceDSE(bEnforceDSE), 
-		  bEnforceNoKDbg(bEnforceNoKDbg), bUseAntiDebugging(bUseAntiDebugging), bCheckIntegrity(bCheckIntegrity), bCheckThreads(bCheckThreads), bCheckHypervisor(bCheckHypervisor), 
-		  bRequireRunAsAdministrator(bRequireRunAsAdministrator), bUsingDriver(bUsingDriver), DriverSignerSubject(DriverSignerSubject), allowedParents(allowedParents), bEnableLogging(bEnableLogging), logFileName(logFileName)
-	{
-		if (Instance != nullptr) //since we can't use a private constructor with ProtectedMemory class, we need to check if the instance is already created
-		{
-			throw runtime_error("The Settings object instance already exists!");
-		}
-
-		Logger::enableLogging = bEnableLogging; //put this line here to be as early as possible.
-		Logger::logFileName = logFileName;
-	}
-	 
-	static Settings* Instance; //singleton-style instance
-
+ 
 private:
 	const wstring KMDriverName = L"UltimateKernelAnticheat"; //optional hybrid approach
 	const wstring KMDriverPath = L".\\UltimateKernelAnticheat.sys";