+++ Obscure/XorStr.hpp
@@ -48,6 +48,16 @@ public:
         }
     }
 
+    std::string decrypt() const
+    {
+        std::string output(N, '\0');
+        for (std::size_t i = 0; i < N; ++i)
+        {
+            output[i] = decrypt_char(encrypted[i], XorKey, i);
+        }
+        return output;
+    }
+
     constexpr int getSize() const { return N; }
 };
 
@@ -79,6 +89,16 @@ public:
         {
             output[i] = decrypt_wchar(encrypted[i], XorKey, i);
         }
+    }
+
+    std::wstring decrypt() const
+    {
+        std::wstring output(N, L'\0');
+        for (std::size_t i = 0; i < N; ++i)
+        {
+            output[i] = decrypt_wchar(encrypted[i], XorKey, i);
+        }
+        return output;
     }
 
     constexpr int getSize() const { return N; }